requirements.txt
pygame==1.9.6
PyOpenGL==3.1.5
PyGLM==1.2.0
pylint==2.5.2
yapf==0.30.0